requirements.txt
pandas
yfinance>=0.2.40
line-bot-sdk